+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Master
 
+* Support conversion of `at_least(0)`
 * Add `-f` shorthand for `--force` option
 
 ## v0.0.9